[B+C] Added FuseStartedEvent and TNTIgniteEvent. Fixes BUKKIT-770
There is currently no event fired when a TNT block is ignited, making it
impossible to link the TNTPrimed entities with the source TNT blocks.

FuseStarteEvent is a generic fuse event for all fused explosives, and should be
extended to TNTMinecart and Creepers (maybe with a new interface for fused
explosives).
TNTIgniteEvent is the specific event for TNT, to be able to get the source
block as well.

CB-1088 - https://github.com/Bukkit/CraftBukkit/pull/1088 - Implementation
